refactor(gcb): address review on cmrel GCB pinning
Follow-up to pinning the GCB cmrel install to the invoking binary's
commit, applying maintainer review feedback:

- Move ref resolution out of cmd/cmrel/cmd (which is one-file-per
  subcommand) into a new pkg/releaseref package, exposing Resolve().
- Handle the !ok from debug.ReadBuildInfo() in Resolve() and pass only
  the *debug.BuildInfo into fromBuildInfo, keeping the resolver a
  single-responsibility, easily testable function.
- Add a --skip-cmrel-pin opt-out to publish so cmrel development against
  a pushed dev branch is not blocked by the fail-closed pin; pinning
  remains the secure default.
- Revert the stage and bootstrap-pgp changes (code and cloudbuild.yaml
  defaults) as those subcommands are effectively unused; only publish is
  pinned.
