+++ gcb/bootstrap-pgp/cloudbuild.yaml
@@ -27,9 +27,7 @@ substitutions:
   ## Required parameters
   _KMS_KEY: ""
   ## Options controlling the version of the release tooling used in the build.
-  ## Must be an immutable commit SHA: cmrel runs here with KMS access to the signing
-  ## key, so a mutable ref (e.g. "master") would execute unpinned code.
-  _RELEASE_REPO_REF: "PINNED_CMREL_COMMIT_SHA_REQUIRED"
+  _RELEASE_REPO_REF: "master"
 
 # we don't specify a machineType manually because we don't need high CPU at all for this
 # job; the default machine type seems to be the lowest-power and cheapest so we leave it
